fixed wrong namespace in Android, updated lock files
