+++ android/app/src/main/java/org/pftp/treemapper/MainApplication.java
@@ -8,6 +8,7 @@ import com.facebook.react.ReactPackage;
 import com.facebook.react.defaults.DefaultNewArchitectureEntryPoint;
 import com.facebook.react.defaults.DefaultReactNativeHost;
 import com.facebook.soloader.SoLoader;
+
 import java.util.List;
 
 public class MainApplication extends Application implements ReactApplication {
